Overhauled message display and added notifications
The message display for normal viewing has been moved to the client-side and is displayed using JS (instead of the old PHP system) but PHP is still used for showing the messages results for a search. Notifications have been also added, showing a popup when a new message is posted from a different device. Also changed the way that the client checks for new messages by keeping the time of the last message and comparing that instead of only checking for new messages from within the last two seconds (to prevent it from missing messages). And added auto smooth scrolling to bottom of page.